Update custom install directions
